Fix the SCIM API
